Move Sign in with Apple responsiblity from AuthVC to UserFirebaseManager and AuthViewModel.
