+++ FindYourOnlys/Model/Firebase/UserFirebaseManager.swift
@@ -7,12 +7,18 @@
 
 import Foundation
 import Firebase
+import AuthenticationServices
+import FirebaseAuth
+import CryptoKit
 
 class UserFirebaseManager {
     
     static let shared = UserFirebaseManager()
     
     private let db = Firestore.firestore()
+    
+    // Unhashed nonce.
+    var currentNonce: String?
     
     
     // swiftlint:disable line_length
@@ -38,6 +44,127 @@ class UserFirebaseManager {
                     friends: ["123"], limitedUsers: ["444"])
     }
     
+    func createAppleIdRequest() -> ASAuthorizationAppleIDRequest {
+        
+        let appleIdProvider = ASAuthorizationAppleIDProvider()
+        
+        let request = appleIdProvider.createRequest()
+        
+        request.requestedScopes = [.fullName, .email]
+        
+        let nonce = randomNonceString()
+        
+        request.nonce = sha256(nonce)
+        
+        currentNonce = nonce
+        
+        return request
+    }
+    
+    // Adapted from https://auth0.com/docs/api-auth/tutorials/nonce#generate-a-cryptographically-random-nonce
+    private func randomNonceString(length: Int = 32) -> String {
+        
+        precondition(length > 0)
+        
+        let charset: [Character] =
+        Array("0123456789ABCDEFGHIJKLMNOPQRSTUVXYZabcdefghijklmnopqrstuvwxyz-._")
+        
+        var result = ""
+        
+        var remainingLength = length
+        
+        while remainingLength > 0 {
+            
+            let randoms: [UInt8] = (0 ..< 16).map { _ in
+                
+                var random: UInt8 = 0
+                
+                let errorCode = SecRandomCopyBytes(kSecRandomDefault, 1, &random)
+                
+                if errorCode != errSecSuccess {
+                    
+                    fatalError(
+                        "Unable to generate nonce. SecRandomCopyBytes failed with OSStatus \(errorCode)"
+                    )
+                }
+                return random
+            }
+            
+            randoms.forEach { random in
+                
+                if remainingLength == 0 {
+                    
+                    return
+                }
+                
+                if random < charset.count {
+                    
+                    result.append(charset[Int(random)])
+                    
+                    remainingLength -= 1
+                }
+            }
+        }
+        
+        return result
+    }
+    
+    func didCompleteWithAuthorization(with authorization: ASAuthorization, completion: @escaping (Error?) -> Void) {
+        
+        if
+            let appleIdCredential = authorization.credential as? ASAuthorizationAppleIDCredential {
+            
+            guard
+                let nonce = currentNonce
+                    
+            else {
+                
+                print("Error, a login callback was recevie, but no request was sent.")
+                
+                return
+            }
+            
+            guard
+                let appleIdToken = appleIdCredential.identityToken
+                    
+            else {
+                print("Can't fetch identity token.")
+                
+                return
+            }
+            
+            guard
+                let idTokenString = String(data: appleIdToken, encoding: .utf8)
+                    
+            else {
+                print("Unable to encode appleIdToken: \(appleIdToken)")
+                
+                return
+            }
+            
+            let credential = OAuthProvider.credential(
+                withProviderID: "apple.com",
+                idToken: idTokenString,
+                rawNonce: nonce
+            )
+            
+            Auth.auth().signIn(with: credential) { authDataResult, error in
+                if
+                    let user = authDataResult?.user {
+                    
+                    print("You have succeed to login, \(user.uid), \(String(describing: user.email))")
+                } else {
+                    
+                    print("Fail to sign in with credential with error: \(String(describing: error))")
+                    
+                    completion(error)
+                }
+            }
+            
+        }
+        
+    }
+    
     func fetchUser(completion: @escaping (Result<[User], Error>) -> Void) {
         
         db.collection(FirebaseCollectionType.user.rawValue)
@@ -55,8 +182,7 @@ class UserFirebaseManager {
                         let user = try document.data(as: User.self)
                         
                         users.append(user)
-                    }
-                    catch {
+                    } catch {
                         
                         completion(.failure(error))
                     }
@@ -64,6 +190,22 @@ class UserFirebaseManager {
                 
                 completion(.success(users))
             }
+    }
+    
+    @available(iOS 13, *)
+    private func sha256(_ input: String) -> String {
+        
+        let inputData = Data(input.utf8)
+        
+        let hashedData = SHA256.hash(data: inputData)
+        
+        let hashString = hashedData.compactMap {
+            
+            String(format: "%02x", $0)
+            
+        }.joined()
+        
+        return hashString
     }
     
     // MARK: - Convert functions
@@ -84,4 +226,5 @@ class UserFirebaseManager {
         
         viewModels.value = UserFirebaseManager.shared.convertUserToViewModels(from: users)
     }
+    
 }
